ref: unnecessary route group removed and Provides.tsx wrapping all providers added
now, the layout.tsx is server components and able to generate metadata and share it with sub pages!
