+++ app/globals.css
@@ -1,7 +1,7 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
- 
+
 @layer base {
   :root {
     --fs--100: clamp(0.5788rem, 0.5509rem + 0.1392vw, 0.64rem);
@@ -15,64 +15,64 @@
     --fs-800: clamp(2.49rem, calc(2.03rem + 2.31vw), 3.82rem);
     --fs-900: clamp(2.99rem, calc(2.37rem + 3.10vw), 4.77rem);
     --fs-1000: clamp(3.58rem, calc(2.76rem + 4.13vw), 5.96rem);
-    
+
     --background: 0 0% 100%;
     --foreground: 222.2 84% 4.9%;
     --card: 0 0% 100%;
     --card-foreground: 222.2 84% 4.9%;
-    
+
     --popover: 0 0% 100%;
     --popover-foreground: 222.2 84% 4.9%;
-    
+
     --primary: 222.2 47.4% 11.2%;
     --primary-foreground: 210 40% 98%;
-    
+
     --secondary: 210 40% 96.1%;
     --secondary-foreground: 222.2 47.4% 11.2%;
-    
+
     --muted: 210 40% 96.1%;
     --muted-foreground: 215.4 16.3% 46.9%;
-    
+
     --accent: 210 40% 96.1%;
     --accent-foreground: 222.2 47.4% 11.2%;
-    
+
     --destructive: 0 84.2% 60.2%;
     --destructive-foreground: 210 40% 98%;
-    
+
     --border: 214.3 31.8% 91.4%;
     --input: 214.3 31.8% 91.4%;
     --ring: 222.2 84% 4.9%;
-    
+
     --radius: 0.5rem;
 
     --color-fg-default:#adbac7;
   }
-  
+
   .dark {
     --background: 222.2 84% 10%;
     --foreground: 210 40% 98%;
- 
+
     --card: 222.2 84% 4.9%;
     --card-foreground: 210 40% 98%;
-    
+
     --popover: 222.2 84% 4.9%;
     --popover-foreground: 210 40% 98%;
-    
+
     --primary: 210 40% 98%;
     --primary-foreground: 222.2 47.4% 11.2%;
-    
+
     --secondary: 217.2 32.6% 17.5%;
     --secondary-foreground: 210 40% 98%;
- 
+
     --muted: 217.2 32.6% 17.5%;
     --muted-foreground: 215 20.2% 65.1%;
- 
+
     --accent: 217.2 32.6% 17.5%;
     --accent-foreground: 210 40% 98%;
- 
+
     --destructive: 0 62.8% 30.6%;
     --destructive-foreground: 210 40% 98%;
- 
+
     --border: 217.2 32.6% 17.5%;
     --input: 217.2 32.6% 17.5%;
     --ring: 212.7 26.8% 83.9%;
@@ -81,6 +81,15 @@
 html{
   scroll-behavior: smooth;
 }
+
+/* It's being auto generated somehow, guess nextUI generates it */
+div[data-overlay-container="true"]{
+  min-height: 100dvh;
+  display: flex;
+  flex-direction: column;
+}
+
+
 .button {
     cursor: pointer;
     font-size: 1.5rem;
